// src/theling-utils/@types/IBase.d.ts
export type UploadFileStatus = 'error' | 'success' | 'done' | 'uploading' | 'removed' | 'waiting';
export interface UploadFile<T = any> {
    uid: string;
    size?: number;
    name: string;
    fileName?: string;
    lastModified?: number;
    lastModifiedDate?: Date;
    url?: string;
    status?: UploadFileStatus;
    percent?: number;
    thumbUrl?: string;
    crossOrigin?: any;
    originFileObj?: any;
    response?: T;
    error?: any;
    linkProps?: any;
    type?: string;
    xhr?: T;
    preview?: string;
}
/**
 * 文件上传的基本对象
 */
export interface IFileObjVoBase extends UploadFile {
    /**
     * 文件id
     */
    fileId: string;
    /**
     * 文件大小
     */
    fileSize: number;
    /**
     * 是否是旧的文件
     */
    oldFile: boolean;
    /**
     * 文件的id，这个是后端生成的，不是文件的uid
     */
    zlFileId: string;
    /**
     * 文件的用途，使用字典：theling_file_use_type
     * @deprecated 请使用 busiScene 代替
     */
    usage?: IFileUpload2UsageType;
    /**
     * 业务场景 -新）
     */
    busiScene?: string;
    cosKey: string;
    /**
     * 带水印的key
     */
    cosWaterKey: string;
    /**
     * 文件的预览图访问地址
     */
    previewUrl: string;
    /**
     * 文件的桶 - cycm会用到
     */
    bucket: string;
    /**
     * 文件的权限 - 0 公开，1 私有，2 登录后可以查看
     */
    permission?: ZlPermissionEnum;
    /**
     * 后缀
     */
    suffix?: string;
    /**
     * 应用的类型
     * @deprecated 请使用 busiScene 代替
     */
    useType?: IFileUpload2UsageType;
    fileUsage?: IFileUpload2UsageType;
    /**
     * 文件关联的id - cycm会用到
     */
    relId?: string;
    /**
     * md5
     */
    md5?: string;
    /**
     * 文件的类型
     * @deprecated 请使用 busiScene 代替
     */
    fileType?: any;
    /**
     * 业务类型 - cycm会用到
     * @deprecated 请使用 busiScene 代替
     */
    busiType?: any;
    /**
     * 创建时间
     */
    creTime?: string;
    /**
     * 给ReactNative用的
     */
    uri?: string;
    /**
     * 文件时长
     */
    videoDuration?: number;
    /**
     * 文件的formdata格式
     */
    formData?: FormData;
}